Revert "Replace ordered ids with hash algo using DJB2"
This reverts commit 3cb953c6c685f194d1ac064b33adcbb737581978.

+++ src/g_game.c
@@ -7071,7 +7071,7 @@ void G_LoadDemoInfo(menudemo_t *pdemo)
 	count = READUINT16(info_p);
 	while (count--)
 	{
-		unsigned long netid;
+		UINT16 netid;
 		char *svalue;
 
 		netid = READUINT16(info_p);
